Use usize/isize and callbacks

+++ src/game/gamedata.rs
@@ -90,7 +90,7 @@ impl Game {
 
     /// returns vector of cards available under a certain cost
     /// hopefully we can combine this and related methods into one generic one
-    pub fn return_avail_cards_ucost(&self, cost: i32) -> Vec<Box<dyn Card>> {
+    pub fn return_avail_cards_ucost(&self, cost: usize) -> Vec<Box<dyn Card>> {
         //TODO: rewrite to not use collect and to use filter() with the lambda passed in
         return self.supply.keys()
             .filter(|a| *self.supply.get(*a).unwrap() > 0)
@@ -101,7 +101,7 @@ impl Game {
 
     /// returns vector of cards available above a certain cost
     /// hopefully we can combine this and related methods into one generic one
-    pub fn return_avail_cards_acost(&self, cost: i32) -> Vec<Box<dyn Card>> {
+    pub fn return_avail_cards_acost(&self, cost: usize) -> Vec<Box<dyn Card>> {
         //TODO: rewrite to not use collect and to use filter() with the lambda passed in
         return self.supply.keys()
             .filter(|a| *self.supply.get(*a).unwrap() > 0)
